Import/Export Agents from Buffer and Parameters
