+++ system_info.txt
@@ -1,0 +1,7 @@
+OS: Linux-5.10.16.3-microsoft-standard-WSL2-x86_64-with-glibc2.35 #1 SMP Fri Apr 2 22:23:49 UTC 2021
+Python: 3.10.6
+Stable-Baselines3: 1.6.2
+PyTorch: 1.13.0+cu117
+GPU Enabled: False
+Numpy: 1.23.5
+Gym: 0.21.0
